refactor: rename package to table - to differentiate from the domain
