+++ data/src/main/java/com/madrid/data/dataSource/local/mappers/GenreMappers.kt
@@ -1,7 +1,7 @@
 package com.madrid.data.dataSource.local.mappers
 
-import com.madrid.data.dataSource.local.entity.MovieGenreEntity
-import com.madrid.data.dataSource.local.entity.SeriesGenreEntity
+import com.madrid.data.dataSource.local.table.MovieGenreEntity
+import com.madrid.data.dataSource.local.table.SeriesGenreEntity
 import com.madrid.data.dataSource.remote.response.genre.MovieGenre
