+++ data/src/main/java/com/madrid/data/dataSource/local/table/relationship/MovieGenreRelation.kt
@@ -1,11 +1,11 @@
-package com.madrid.data.dataSource.local.entity.relationship
+package com.madrid.data.dataSource.local.table.relationship
 
 import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.Junction
 import androidx.room.Relation
-import com.madrid.data.dataSource.local.entity.MovieGenreEntity
-import com.madrid.data.dataSource.local.entity.MovieEntity
+import com.madrid.data.dataSource.local.table.MovieGenreEntity
+import com.madrid.data.dataSource.local.table.MovieEntity
 
 @Entity(tableName = "MovieGenreCrossRef", primaryKeys = ["genreId", "movieId"])
 data class MovieGenreCrossRef(
